front end peliculas con logie

+++ src/main/java/es/uah/cFilmsActores/controller/UsersController.java
@@ -1,9 +1,10 @@
 package es.uah.cFilmsActores.controller;
 
 
-import es.uah.cFilmsActores.model.Actor;
+import es.uah.cFilmsActores.model.Rol;
 import es.uah.cFilmsActores.model.User;
 import es.uah.cFilmsActores.paginator.PageRender;
+import es.uah.cFilmsActores.service.IRolesService;
 import es.uah.cFilmsActores.service.IUsersService;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +16,8 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import java.util.Arrays;
+import java.util.List;
 import java.util.logging.Logger;
 
 @Controller
@@ -23,6 +26,10 @@ public class UsersController {
 
     @Autowired
     IUsersService usersService;
+
+
+    @Autowired
+    IRolesService rolesService;
 
 
 @GetMapping (value = "/all")
@@ -45,7 +52,7 @@ public String usersList (Model model, @RequestParam(name = "page", defaultValue 
 }
 
 
-@GetMapping ("/search")
+    @GetMapping ("/search")
     public String search (Model model) {
     return "users/searchUser";
 }
@@ -57,7 +64,7 @@ public String usersList (Model model, @RequestParam(name = "page", defaultValue 
     return "formUsers";
             }
 
-            @GetMapping ("/username")
+    @GetMapping ("/username")
     public String findUserByUsername (Model model, @RequestParam("username") String username) {
     User user = usersService.findUserByUsername(username);
     model.addAttribute("username", "listado de usuarios por username");
@@ -82,20 +89,50 @@ public String usersList (Model model, @RequestParam(name = "page", defaultValue 
     attributes.addFlashAttribute("msg", "los datos del usuario fueron guardados !");
     return "redirect:/users/all";
     }
+    @GetMapping("/registrarme")
+    public String nuevoRegistro(Model model) {
+        model.addAttribute("title", "Nuevo registro");
+        User user = new User();
+        model.addAttribute("user", user);
+        return "/registro";
+    }
+    @PostMapping("/registrar")
+    public String registro(Model model, User user, RedirectAttributes attributes) {
+        //si existe un usuario con el mismo correo no lo guardamos
+        if (usersService.findUserByEmail(user.getEmail())!=null) {
+            attributes.addFlashAttribute("msga", "Error al guardar, ya existe el correo!");
+            return "redirect:/login";
+        }
+        user.setEnable(true);
+        Rol rol = rolesService.findRolById(1);
+        user.setRoles(Arrays.asList(rol));
+        usersService.saveUser(user);
+        attributes.addFlashAttribute("msg", "Los datos del registro fueron guardados!");
+        return "redirect:/login";
+    }
+
 
 
     @GetMapping("edit/{id}")
     public String editUser(Model model, @PathVariable("id") Integer id) {
     User user = usersService.findUserById(id);
-    model.addAttribute("titule", "Editar Usuario");
+    model.addAttribute("title", "Editar Usuario");
     model.addAttribute("user", user);
+    List<Rol> roles = rolesService.findAll();
+    model.addAttribute("allRoles", roles);
     return "users/formUsers";
     }
 
     @GetMapping ("/delete/{id}")
     public String deleteUser(Model model, @PathVariable("id")Integer id, RedirectAttributes attributes){
+
+    User user = usersService.findUserById(id);
+    if ( user != null) {
     usersService.deleteUser(id);
     attributes.addFlashAttribute("msg", "los datos del usuario fueron borrados!");
+    } else {
+        attributes.addFlashAttribute("msg", "Usuario no encontrado!");
+    }
     return "redirect:/users/allUsers";
     }
